% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BottomUpMerge.R
\name{BottomUpMerge}
\alias{BottomUpMerge}
\title{Generate clusterings by iteratively merging two nearest clusters for a Seurat object}
\usage{
BottomUpMerge(sobj, k.max, npc, random.seed)
}
\arguments{
\item{sobj}{Seurat object}

\item{k.max}{the maximal number of clusters}

\item{npc}{the number of top principal components used for Seurat SNN clustering}

\item{random.seed}{the random seed set for Seurat SNN clustering}
}
\description{
This function generates clusterings by iteratively merging two nearest clusters for a Seurat object. It generates an
initial clustering with the number of clusters (k) >= k.max using Seurat SNN clustering. Then iteratively merge
two clusters with the nearest centers to generate the clustering with k-1 clusters. Finally, it returns k.max
clusterings with k = 1 to k.max.
}
\examples{
BottomUpMerge(sobj, k.max, npc, random.seed)
}
\keyword{IKAP}
