% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeMarkers.R
\name{ComputeMarkers}
\alias{ComputeMarkers}
\title{Compute upregulated marker genes for each cell group in the Seurat object}
\usage{
ComputeMarkers(sobj, gap.gain, candidates, out.dir)
}
\arguments{
\item{sobj}{the Seurat object}

\item{gap.gain}{a numeric matrix for the increase of gap statistic from k-1 to k. Columns are
the number of clusters (k = 2 to k.max) and rows are the number of top principal components (nPC).
For example, the first column is the increase of gap statistic from k=1 to k=2 for each nPC (row).}

\item{candidates}{the list of candidates returned from SelectCandidate().}

\item{out.dir}{the path for output directory}
}
\description{
This function computes the list of upregulated marker genes for each cell group in the Seurat object 
using Seurat::FindAllMarkers. It also computes the area under the ROC curve (AUROC) for each marker gene.
Finally it saves all relevant in Excel sheets.
}
\examples{
ComputeMarkers(sobj, gap.gain, candidates, out.dir)
}
\keyword{IKAP}
