% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionTree.R
\name{DecisionTree}
\alias{DecisionTree}
\title{Build decision trees for each cell group in the candidate sets}
\usage{
DecisionTree(sobj, markers, out.dir, plot.decision.tree)
}
\arguments{
\item{sobj}{the Seurat object}

\item{markers}{a list of marker gene tables for candidate sets. This is usually the list returned from
ComputeMarkers().}

\item{out.dir}{the path for output directory}

\item{plot.decision.tree}{whether to plot decision trees?}
}
\description{
This function builds decision trees for each cell group each candidate set using the R package rpart.
The overall classification error is summarized for each candidate set of cell groups.
}
\examples{
DecisionTree(sobj, markers, out.dir, plot.decision.tree)
}
\keyword{IKAP}
