% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedLogW.R
\name{ExpectedLogW}
\alias{ExpectedLogW}
\title{Compute log expected sum of pairwise distances for a given grouping}
\usage{
ExpectedLogW(pc.ranges, one.clustering)
}
\arguments{
\item{pc.ranges}{a numeric matrix: the minimum (1st row) and the maximum (2nd row) for each dimension (column)}

\item{one.clustering}{a vector indicating the clustering membership for each data point}
}
\description{
This function computes log expected sum of pairwise distances for a given grouping. It assumes data points are
uniformly distributed within a range of values in every dimension. It first computes the expected sum of
pairwise distances for each group (cluster) and then takes log for the sum of expected sums acros all groups.
}
\examples{
ExpectedLogW(pc.ranges, one.clustering)
}
\keyword{IKAP}
