% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapStatistic.R
\name{GapStatistic}
\alias{GapStatistic}
\title{Compute gap statistic}
\usage{
GapStatistic(x, clusterings)
}
\arguments{
\item{x}{a numeric matrix: Columns are samples and rows are dimensions.}

\item{clusterings}{a list of vectors, each of which indicates the clustering label for each data point}
}
\description{
This function computes gap statistics for K clusterings with the number of clusters = 1 to K.
}
\examples{
GapStatistic(x, clusterings)
}
\keyword{IKAP}
