% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IKAP.R
\name{IKAP}
\alias{IKAP}
\title{IKAP: Identifying K mAjor cell Population group in single cell analysis}
\usage{
IKAP(sobj, pcs = NA, pc.range = 20, k.max = NA, r.kmax.est = 1.5,
  out.dir = "./IKAP", scale.data = TRUE, confounders = c("nUMI",
  "percent.mito"), plot.decision.tree = TRUE, random.seed = 0)
}
\arguments{
\item{sobj}{a Seurat object with cell expression normalized}

\item{pcs}{the list of principal components used for clustering. default is NA (to be determined by IKAP; recommended)}

\item{pc.range}{the range of nPCs. default is 20}

\item{k.max}{the maximal number of clusters. default is NA (to be determined by IKAP; recommended)}

\item{r.kmax.est}{resolution for IKAP determining k.max using Seurat SNN clustering. default is 1.5}

\item{out.dir}{the path for output directory}

\item{scale.data}{whether scale the data using Seurat::ScaleData. default is TRUE (recommended)}

\item{confounders}{a vector of confounders that need to be regressed out in Seurat::ScaleData.
default is c('nUMI','percent.mito') (see Seurat tutorial: https://satijalab.org/seurat/pbmc3k_tutorial.html)}

\item{plot.decision.tree}{whether to plot decision trees that classify the cell groups. default is TRUE}

\item{random.seed}{random seed}
}
\description{
IKAP identifies candidate set(s) of major cell groups using the single cell analysis R package
Seurat by evaluating sets of possible cell groups generated using different parameters in Seurat
SNN clustering (i.e. resulotion r and the number of top principal components (nPC)). The results
(tables and plots) are saved in the output directory. A Seurat object is returned with all sets of
evaluated cell groups saved in the metadata data frame.
}
\examples{
sobj.new <- IKAP(sobj, out.dir = "./IKAP")

saveRDS(sobj.new, file = "./IKAP/sobj.new.rds")
}
\keyword{IKAP}
