% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestCluster.R
\name{NearestCluster}
\alias{NearestCluster}
\title{Report the two clusters with the nearest centers}
\usage{
NearestCluster(x, clustering)
}
\arguments{
\item{x}{a numeric matrix. Columns are samples and rows are dimensions.}

\item{clustering}{a vector indicating the clustering label for each data point.}
}
\description{
This function reports the two clusters with the nearest centers.
}
\examples{
NearestCluster(x, clustering)
}
\keyword{IKAP}
