% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservedLogW.R
\name{ObservedLogW}
\alias{ObservedLogW}
\title{Compute log observed sum of pairwise distances for a given grouping}
\usage{
ObservedLogW(X, one.clustering)
}
\arguments{
\item{X}{a numeric matrix. Columns are samples and rows are dimensions.}

\item{one.clustering}{a vector indicating the clustering membership for each data point}
}
\description{
This function computes log observed sum of pairwise distances for a given grouping. It first computes the sum of
pairwise distances for each group (cluster) and then takes log for the sum of sums acros all groups.
}
\examples{
ObservedLogW(pc.ranges, one.clustering)
}
\keyword{IKAP}
