% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSummary.R
\name{PlotSummary}
\alias{PlotSummary}
\title{Plot the IKAP result}
\usage{
PlotSummary(gap.gain, summary.rpart, markers.all, out.dir)
}
\arguments{
\item{gap.gain}{a numeric matrix for the increase of gap statistic from k-1 to k. Columns are
the number of clusters (k = 2 to k.max) and rows are the number of top principal components (nPC).
For example, the first column is the increase of gap statistic from k=1 to k=2 for each nPC (row).}

\item{summary.rpart}{a numberic matrix with overall classification error summarized for each n split (column)
and each candidate set (row). This is usually the output return from DecisionTree().}

\item{markers.all}{a list of marker gene tables for candidate sets. This is usually the list returned from
ComputeMarkers().}

\item{out.dir}{the path for output directory}
}
\description{
This function plots a heat map indicating the increase of gap statistics as the number of cluster (k)
increases by varying the number of top principal components (nPC) used. The best set and alternative
sets of cell groups are indicated by 'B' and 'X' respectively. The best set is determined based on
classification error summarized in DecisionTree().
}
\examples{
PlotSummary(gap.gain, summary.rpart, markers.all, out.dir)
}
\keyword{IKAP}
