% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectCandidate.R
\name{SelectCandidate}
\alias{SelectCandidate}
\title{Select the candidate sets of cell groups}
\usage{
SelectCandidate(gap.gain)
}
\arguments{
\item{gap.gain}{a numeric matrix for the increase of gap statistic from k-1 to k. Columns are
the number of clusters (k = 2 to k.max) and rows are the number of top principal components (nPC).
For example, the first column is the increase of gap statistic from k=1 to k=2 for each nPC (row).}
}
\description{
This function selects the candidate sets of cell groups based on the changes of gap statistics.
}
\examples{
SelectCandidate(gap.gain)
}
\keyword{IKAP}
